Move CompanyTable
