+++ src/frontend/src/pages/purchasing/PurchasingIndex.tsx
@@ -9,7 +9,7 @@ import { useMemo } from 'react';
 
 import { PageDetail } from '../../components/nav/PageDetail';
 import { PanelGroup } from '../../components/nav/PanelGroup';
-import { CompanyTable } from '../../components/tables/general/CompanyTable';
+import { CompanyTable } from '../../components/tables/company/CompanyTable';
 import { PurchaseOrderTable } from '../../components/tables/purchasing/PurchaseOrderTable';
 
 export default function PurchasingIndex() {
